+++ boot/index.ts
@@ -6,6 +6,7 @@ import { CourseEnrollmentStorage } from "@/services/course-enrollments-storage";
 import { CoursePaymentStorage } from "@/services/course-payments-storage";
 import { ContactStorage } from "@/services/contacts-storage";
 import { PasswordResetAttemptStorage } from "@/services/password-reset-attempts-storage";
+import { RabbiBioStorage } from "@/services/rabbi-bio-storage";
 
 export async function Boot() {
   await UserStorage.ensureIndexes();
@@ -15,5 +16,6 @@ export async function Boot() {
   await CoursePaymentStorage.ensureIndexes();
   await ContactStorage.ensureIndexes();
   await PasswordResetAttemptStorage.ensureIndexes();
+  await RabbiBioStorage.ensureIndexes();
   await OperatorStorage.ensureDefaultOperator();
 }